+++ internal/resources/advancedusersearches/resource.go
@@ -16,10 +16,10 @@ func ResourceJamfProAdvancedUserSearches() *schema.Resource {
 		UpdateContext: update,
 		DeleteContext: delete,
 		Timeouts: &schema.ResourceTimeout{
-			Create: schema.DefaultTimeout(120 * time.Second),
-			Read:   schema.DefaultTimeout(15 * time.Second),
-			Update: schema.DefaultTimeout(30 * time.Second),
-			Delete: schema.DefaultTimeout(15 * time.Second),
+			Create: schema.DefaultTimeout(70 * time.Second),
+			Read:   schema.DefaultTimeout(70 * time.Second),
+			Update: schema.DefaultTimeout(70 * time.Second),
+			Delete: schema.DefaultTimeout(70 * time.Second),
 		},
 		Importer: &schema.ResourceImporter{
 			StateContext: schema.ImportStatePassthroughContext,