+++ internal/resources/policies/resource.go
@@ -21,10 +21,10 @@ func ResourceJamfProPolicies() *schema.Resource {
 			StateContext: schema.ImportStatePassthroughContext,
 		},
 		Timeouts: &schema.ResourceTimeout{
-			Create: schema.DefaultTimeout(120 * time.Second),
-			Read:   schema.DefaultTimeout(15 * time.Second),
-			Update: schema.DefaultTimeout(30 * time.Second),
-			Delete: schema.DefaultTimeout(15 * time.Second),
+			Create: schema.DefaultTimeout(70 * time.Second),
+			Read:   schema.DefaultTimeout(70 * time.Second),
+			Update: schema.DefaultTimeout(70 * time.Second),
+			Delete: schema.DefaultTimeout(70 * time.Second),
 		},
 		SchemaVersion: 1,
 		StateUpgraders: []schema.StateUpgrader{